implemented Google Ads
add trash explosion damage multiply

+++ Assets/Scripts/Trash/TrashTypeSO.cs
@@ -5,9 +5,9 @@ public class TrashTypeSO : ScriptableObject
 {
     [SerializeField] private string _typeName;
     [SerializeField] private int _scoreValue;
-    [SerializeField] private int _polutionValue;
+    [SerializeField] private float _polutionValue;
 
     public string TypeName => _typeName;
     public int ScoreValue => _scoreValue; 
-    public int PolutionValue => _polutionValue; 
+    public float PolutionValue => _polutionValue; 
 }
